Harden hydration smoke login to match the resilient browse helper.
The hydration check was racing the member list before session/members settled; reuse the shared login helper so CI does not flake on cold starts.
